Rename ChatServer.py to server.py
